check for expiry when atom gets updated
